Start a new include file for the add-on logic
