+++ tests/test-admin.php
@@ -131,68 +131,6 @@ class AdminTest extends TestCase {
 		$this->assertContains( get_admin_url( null, 'admin.php?page=wp101&configure=1' ), $actions[0] );
 	}
 
-	public function test_check_plugins() {
-		update_option( 'active_plugins', [
-			'some-plugin/some-plugin.php',
-			'another-plugin/another-plugin.php',
-		] );
-
-		$api = $this->mock_api();
-		$api->shouldReceive( 'get_addons' )
-			->once()
-			->andReturn([
-				'addons' => [
-					[
-						'title'                  => 'Learning Some Plugin',
-						'url'                    => 'https://wp101plugin.com/series/some-plugin',
-						'includedInSubscription' => false,
-						'restrictions'           => [
-							'plugins' => [
-								'some-plugin/some-plugin.php',
-							],
-						],
-					],
-				],
-			]);
-
-		Admin\check_plugins();
-
-		$this->assertEquals([
-			[
-				'title'  => 'Learning Some Plugin',
-				'url'    => 'https://wp101plugin.com/series/some-plugin',
-				'plugin' => 'some-plugin/some-plugin.php',
-			],
-		], get_option( 'wp101-available-series', [] ) );
-	}
-
-	public function test_check_plugins_excludes_addons_included_in_subscription() {
-		update_option( 'active_plugins', [
-			'some-plugin/some-plugin.php',
-		] );
-
-		$api = $this->mock_api();
-		$api->shouldReceive( 'get_addons' )
-			->andReturn([
-				'addons' => [
-					[
-						'title'                  => 'Learning Some Plugin',
-						'url'                    => 'https://wp101plugin.com/series/some-plugin',
-						'includedInSubscription' => true,
-						'restrictions'           => [
-							'plugins' => [
-								'some-plugin/some-plugin.php',
-							],
-						],
-					],
-				],
-			]);
-
-		Admin\check_plugins();
-
-		$this->assertEmpty( get_option( 'wp101-available-series', [] ) );
-	}
-
 	/**
 	 * Retrieve the WP101 menu node(s) visible for the current user.
 	 *
